refactor(harness): move MockModel impl out of providers/mod.rs
providers/mod.rs carried ~275 lines of MockModel constructors, the
ChatModel impl, and token-estimation helpers alongside the module's
own wiring/doc comment. Moved into mock.rs, leaving mod.rs with just
module declarations and shared imports. MockModel's type definition
already lived in types.rs per convention. No behavior change.
